refactor(toolbox): extract shared backtracking assignment matcher
Extract the private assignable function from useOvermasteryPredictor into
a new shared matching.ts module. The generic signature enables reuse by
wrightstone matching without code duplication.

Behavior is identical — all 31 existing tests pass.
